% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{as.matrix.sparseZ2}
\alias{as.matrix.sparseZ2}
\title{Convert Sparse Matrix to Regular Matrix.}
\usage{
\method{as.matrix}{sparseZ2}(x, ...)
}
\arguments{
\item{x}{A \code{sparseZ2} matrix}

\item{...}{ignored.}
}
\value{
A regular matrix of 0s and 1s
}
\description{
Convert a sparse matrix to a regular matrix.
}
