% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{dim.sparseZ2}
\alias{dim.sparseZ2}
\title{Dimensions of a Sparse Matrix}
\usage{
\method{dim}{sparseZ2}(x)
}
\arguments{
\item{x}{A \code{sparseZ2} matrix}
}
\value{
A vector of length 2 giving the number of rows and columns.
}
\description{
Get the dimensions of a \code{sparseZ2} matrix
}
