% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{gaussReduceZ2}
\alias{gaussReduceZ2}
\title{Sparse Gaussian Reduction over Z2}
\usage{
gaussReduceZ2(A1, b1, A2, b2)
}
\arguments{
\item{A1}{The reduced sparse matrix from the first system}

\item{b1}{The reduced right-hand side from the first system}

\item{A2}{The sparse matrix to be reduced}

\item{b2}{The right-hand side to be reduced}
}
\value{
A list containing the reduced A2 and b2
}
\description{
Reduce a linear system using a previously reduced system
}
\details{
Given a system A1 x1 = b1 that has been reduced with \code{\link[=gaussJordanZ2]{gaussJordanZ2()}} or
\code{\link[=gaussZ2]{gaussZ2()}}, reduce the system A2 x2 = b2 using the reduced form of A1.
}
\examples{
# Create a sparse matrix
A <- sparseZ2(4, 4)
A[1,c(1,3,4)] <- 1
A[2,c(2,3)] <- 1
A[3,c(1,2,4)] <- 1
A[4,c(1,4)] <- 1

b <- c(1, 0, 1, 1)

# Split the system
A1 <- sparseZ2(2, 4)
A1[1,c(1,3,4)] <- 1
A1[2,c(2,3)] <- 1
b1 <- b[1:2]

A2 <- sparseZ2(2, 4)
A2[1,c(1,2,4)] <- 1
A2[2,c(1,4)] <- 1
b2 <- b[3:4]

# Reduce A1
sys1 <- gaussZ2(A1, b1)
sys1

# Reduce A2 using A1
sys2 <- gaussReduceZ2(sys1$A, sys1$b, A2, b2)
sys2

# Further reduce A2
sys2 <- gaussZ2(sys2$A, sys2$b)
sys2

}
