% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{gaussZ2A}
\alias{gaussZ2A}
\title{Sparse Gaussian Elimination over Z2}
\usage{
gaussZ2A(A, b)
}
\arguments{
\item{A}{A \code{sparseZ2} matrix}

\item{b}{A vector over the field Z2}
}
\value{
A list containing the reduced matrix and right hand side.
}
\description{
Calculates the reduced form of the sparse linear system A x = b by
Gaussian elimination over the field Z2.
}
\details{
This implementation does no pivoting so the order of the rows and
columns is preserved.
}
\examples{
# Create a sparse matrix
A <- sparseZ2(4, 4)
A[1,c(1,3,4)] <- 1
A[2,c(2,3)] <- 1
A[3,c(1,2,4)] <- 1
A[4,c(1,4)] <- 1

b <- c(1, 0, 1, 1)

# Perform Gaussian elimination
sys <- gaussZ2(A, b)
sys

}
