% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{jordanReduceZ2}
\alias{jordanReduceZ2}
\title{Sparse Gauss-Jordan Elimination over Z2}
\usage{
jordanReduceZ2(A, b)
}
\arguments{
\item{A}{A sparse matrix over the field Z2}

\item{b}{A vector over the field Z2}
}
\value{
A list containing the reduced matrix and right hand side
}
\description{
Reduce a linear system in row echelon form to reduced row-echelon form.
}
\details{
Given a Gaussian reduced system computed by \code{\link[=gaussJordanZ2]{gaussJordanZ2()}}
compute the full Gauss-Jordan reduction.
}
\examples{
# Create a sparse matrix
A <- sparseZ2(4, 4)
A[1,c(1,3,4)] <- 1
A[2,c(2,3)] <- 1
A[3,c(1,2,4)] <- 1
A[4,c(1,4)] <- 1

b <- c(1, 0, 1, 1)

# Reduce to row echelon form by Gaussian elimination
sys <- gaussJordanZ2(A, b, jordan=FALSE)
sys

# Reduce to reduced row echelon form
sys <- jordanReduceZ2(sys$A,sys$b)
sys

# Reduce to reduced row echelon form by Gauss-Jordan elimination
sys <- gaussJordanZ2(A, b, jordan=TRUE)
sys

}
