% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{multAb}
\alias{multAb}
\title{Sparse Matrix Vector Multiplication}
\usage{
multAb(A, b)
}
\arguments{
\item{A}{A \code{sparseZ2} matrix}

\item{b}{A vector over Z2}
}
\value{
A vector over Z2
}
\description{
Multiply a \code{sparseZ2} matrix by a vector over the field Z2
}
