% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{nullSpace}
\alias{nullSpace}
\title{Null Space of a Sparse Matrix}
\usage{
nullSpace(A)
}
\arguments{
\item{A}{A sparse matrix over Z2}
}
\value{
A sparse matrix with rows that form a basis for the null space of A
}
\description{
Compute the null space of a sparse matrix over Z2
}
\details{
Given a reduced matrix A and vector b that are consistent, \code{nullSpace}
returns a sparse matrix with rows that are the non-trivial elements of a
basis for the null space of A.
}
\examples{
# Create a sparse matrix
A <- sparseZ2(4, 4)
A[1,c(1,3,4)] <- 1
A[2,c(2,3)] <- 1
A[3,c(1,2,4)] <- 1
A[4,c(1,4)] <- 1

b <- c(1, 0, 1, 1)

## Find solutions
sys <- gaussJordanZ2(A, b, jordan = FALSE)
consistentSystem(sys$A,sys$B)
solution(sys$A,sys$b)
nullSpace(sys$A)
}
