% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{print.sparseZ2}
\alias{print.sparseZ2}
\title{Print method for sparse matrices over Z2}
\usage{
\method{print}{sparseZ2}(x, ...)
}
\arguments{
\item{x}{A \code{sparseZ2} matrix}

\item{...}{Additional arguments passed to print}
}
\description{
Print a \code{sparseZ2} matrix in standard matrix format
}
