% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{rbind.sparseZ2}
\alias{rbind.sparseZ2}
\title{Combine \code{sparseZ2} Matrices by Rows}
\usage{
\method{rbind}{sparseZ2}(..., deparse.level = 1)
}
\arguments{
\item{...}{\code{sparseZ2} matrices to be combined}

\item{deparse.level}{Not used}
}
\value{
A sparse matrix
}
\description{
Combine \code{sparseZ2} matrices by rows
}
