% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{reduce_row}
\alias{reduce_row}
\title{Row Reduction in Z2}
\usage{
reduce_row(r1, r2)
}
\arguments{
\item{r1}{sorted integer vector}

\item{r2}{sorted integer vector}
}
\value{
a sorted integer vector representing the reduced row.
}
\description{
Reduce two rows of a sparse matrix over the field Z2.
}
\details{
Each row of the sparse matrix is represented by an integer vector
the elements of which are the (sorted) column numbers of the
non-zero elements of the row. This function takes sorted integer
vectors representing two rows of a matrix and computes the sorted
integer vector corresponding to sum of the rows in Z2.
}
