% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{solution}
\alias{solution}
\alias{consistentSystem}
\alias{determinedVars}
\alias{undeterminedVars}
\title{Solutions of Linear Systems}
\usage{
solution(A, b)

consistentSystem(A, b)

determinedVars(A)

undeterminedVars(A, all = FALSE)
}
\arguments{
\item{A}{A reduced sparse matrix over Z2}

\item{b}{A reduced vector over Z2}

\item{all}{Include all variables or only those that appear in the matrix.}
}
\value{
TRUE if the system is consistent, FALSE otherwise
}
\description{
Find the solutions of a (possibly underdetermined) linear system over Z2.
}
\details{
Given a reduced matrix A and vector b, \code{consistentSystem} checks if the
system is consistent and has at least one solution.  If the system is
consistent, \code{determinedVars} returns the variables that are determined by
the system, and \code{undeterminedVars} returns the variables that are
undetermined.  When \code{all=FALSE}, \code{undeterminedVars} returns only variables
that occur in the matrix.

Given a Gauss-Jordan reduced matrix A and vector b that is consistent,
\code{solution} returns a one possible solution, with any undetermined variables
set to 0.
}
\examples{
# Create a sparse matrix
A <- sparseZ2(4, 4)
A[1,c(1,3,4)] <- 1
A[2,c(2,3)] <- 1
A[3,c(1,2,4)] <- 1
A[4,c(1,4)] <- 1

b <- c(1, 0, 1, 1)

## Determine solutions
sys <- gaussJordanZ2(A, b, jordan = FALSE)
consistentSystem(sys$A,sys$B)
undeterminedVars(sys$A)

sys <- jordanReduceZ2(sys$A, sys$b)
solution(sys$A, sys$b)

}
