% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2-package.R
\docType{package}
\name{sparseZ2-package}
\alias{sparseZ2-package}
\title{sparseZ2: Sparse matrices over Z2}
\description{
Facilitites for solving sparse linear systems over the field Z2.
}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

}
\keyword{internal}
