% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{sparseZ2}
\alias{sparseZ2}
\title{Create Sparse Matrix over Z2}
\usage{
sparseZ2(nrow, ncol = NA)
}
\arguments{
\item{nrow}{Number of rows in the matrix}

\item{ncol}{Number of columns in the matrix}
}
\value{
A \code{sparseZ2} matrix with all elements 0
}
\description{
Create an empty sparse matrix over the field Z2
}
\details{
The matrix is represented in a sparse format as a list of rows, where
each row is represented by as an integer vector storing the (sorted)
column numbers of the non-zero elements of the row.

It is not necessary to specify the number of columns at the time the
matrix is created, it will be inferred from the maximum non-zero entry.
}
