% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{[.sparseZ2}
\alias{[.sparseZ2}
\title{Extract Elements of a Sparse Matrix.}
\usage{
\method{[}{sparseZ2}(x, i, j)
}
\arguments{
\item{x}{A \code{sparseZ2} matrix}

\item{i}{Row index or vector of row indices}

\item{j}{Column index or vector of column indices}
}
\value{
The value(s) at the specified position(s) (0 or 1)
}
\description{
Extract elements from a sparse matrix over the field Z2
}
\details{
Extract elements from a \code{sparseZ2} matrix.  It is an error for the
column index to be missing.
}
