% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{[<-.sparseZ2}
\alias{[<-.sparseZ2}
\title{Set Elements of a Sparse Matrix.}
\usage{
\method{[}{sparseZ2}(x, i, j) <- value
}
\arguments{
\item{x}{A \code{sparseZ2} matrix}

\item{i}{Row index or vector of row indices}

\item{j}{Column index or vector of column indices}

\item{value}{Value(s) to set (0 or 1)}
}
\value{
Updated sparse matrix
}
\description{
Set elements in a \code{sparseZ2} matrix.
}
\details{
Set elements of a \code{sparseZ2} matrix, with any non-zero values treated as 1.
It is an error for the column index to be missing.
}
