% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseZ2.R
\name{transpose}
\alias{transpose}
\title{Transpose a Sparse Matrix}
\usage{
transpose(A)
}
\arguments{
\item{A}{A \code{sparseZ2} matrix}
}
\value{
The transposed \code{sparseZ2} matrix
}
\description{
Transpose a \code{sparseZ2} matrix
}
